move data_utils to typescript
