+++ src/js/main/util/ui_utils.ts
@@ -1,0 +1,18 @@
+
+const utils = {
+    // set visibility of the "loading..." layer that blocks interaction with any elements
+    setLoadingOverlayVisible: (visible: boolean) => {
+        const loadingScreen = document.getElementById('loading_screen');
+        loadingScreen.classList.remove('hide', 'flex');
+        if (visible) {
+            loadingScreen.classList.add('flex');
+        } else {
+            loadingScreen.classList.add('hide');
+        }
+    }
+};
+
+export default {
+    hideLoadingOverlay: () => utils.setLoadingOverlayVisible(false),
+    showLoadingOverlay: () => utils.setLoadingOverlayVisible(true)
+};
